+++ lib/mmdb2_decoder.ex
@@ -33,13 +33,14 @@ defmodule MMDB2Decoder do
   alias MMDB2Decoder.LookupTree
   alias MMDB2Decoder.Metadata
 
-  @type decode_option :: term
+  @type decode_option :: {:float_precision, Float.precision_range()}
   @type decode_options :: [decode_option]
   @type decoded_value :: :cache | :end | binary | boolean | list | map | number
   @type lookup_result :: {:ok, decoded_value} | {:error, term}
   @type parse_result :: {:ok, Metadata.t(), binary, binary} | {:error, term}
 
-  @default_options []
+  @doc false
+  def default_options, do: [float_precision: 4]
 
   @doc """
   Looks up the data associated with an IP tuple.
@@ -65,7 +66,7 @@ defmodule MMDB2Decoder do
   """
   @spec lookup(:inet.ip_address(), Metadata.t(), binary, binary, decode_options) :: lookup_result
   def lookup(ip, meta, tree, data, options \\ []) do
-    options = Keyword.merge(@default_options, options)
+    options = Keyword.merge(default_options(), options)
 
     case LookupTree.locate(ip, meta, tree) do
       {:error, _} = error -> error